requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==0.12.2
Flask-Cors==3.0.3
future==0.17.1
gym==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.14.1
Pillow==6.0.0
progressbar2==3.37.0
pyglet==1.3.2
python-utils==2.3.0
requests==2.21.0
scikit-learn==0.20.3
scipy==1.0.0
six==1.12.0
sklearn==0.0
urllib3==1.24.2
Werkzeug==0.15.2
